refactor(io-engine/rpc): split the version handling in separate modules
Adds new traits for the pool/replica/nexus/host api's which are implemented by each
version handler rpc.
This should make it much easier to add newer versions in the future.
Also split the node/pool wrapper in separate modules.
