+++ control-plane/agents/src/bin/core/pool/wrapper.rs
@@ -1,0 +1,145 @@
+use common_lib::types::v0::transport::{PoolState, PoolStatus, Protocol, Replica, ReplicaId};
+
+use std::{cmp::Ordering, ops::Deref};
+
+/// Wrapper over the message bus `Pool` which includes all the replicas
+/// and Ord traits to aid pool selection for volume replicas
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub(crate) struct PoolWrapper {
+    state: PoolState,
+    replicas: Vec<Replica>,
+}
+
+impl Deref for PoolWrapper {
+    type Target = PoolState;
+    fn deref(&self) -> &Self::Target {
+        &self.state
+    }
+}
+
+impl PoolWrapper {
+    /// New Pool wrapper with the pool and replicas.
+    pub(crate) fn new(pool: PoolState, replicas: Vec<Replica>) -> Self {
+        Self {
+            state: pool,
+            replicas,
+        }
+    }
+
+    /// Get all the replicas.
+    pub(crate) fn replicas(&self) -> Vec<Replica> {
+        self.replicas.clone()
+    }
+    /// Get the specified replica.
+    pub(crate) fn replica(&self, replica: &ReplicaId) -> Option<&Replica> {
+        self.replicas.iter().find(|r| &r.uuid == replica)
+    }
+    /// Get the state.
+    pub(crate) fn state(&self) -> &PoolState {
+        &self.state
+    }
+
+    /// Get the free space.
+    pub(crate) fn free_space(&self) -> u64 {
+        if self.state.capacity >= self.state.used {
+            self.state.capacity - self.state.used
+        } else {
+            // odd, let's report no free space available
+            tracing::error!(
+                "Pool '{}' has a capacity of '{} B' but is using '{} B'",
+                self.state.id,
+                self.state.capacity,
+                self.state.used
+            );
+            0
+        }
+    }
+
+    /// Set pool state as unknown.
+    #[allow(dead_code)]
+    pub(crate) fn set_unknown(&mut self) {
+        self.state.status = PoolStatus::Unknown;
+    }
+
+    /// Add replica to list.
+    #[allow(dead_code)]
+    pub(crate) fn add_replica(&mut self, replica: &Replica) {
+        self.replicas.push(replica.clone())
+    }
+    /// Remove replica from list.
+    #[allow(dead_code)]
+    pub(crate) fn remove_replica(&mut self, uuid: &ReplicaId) {
+        self.replicas.retain(|replica| &replica.uuid != uuid)
+    }
+    /// Update replica from list.
+    #[allow(dead_code)]
+    pub(crate) fn update_replica(&mut self, uuid: &ReplicaId, share: &Protocol, uri: &str) {
+        if let Some(replica) = self
+            .replicas
+            .iter_mut()
+            .find(|replica| &replica.uuid == uuid)
+        {
+            replica.share = *share;
+            replica.uri = uri.to_string();
+        }
+    }
+}
+
+impl From<PoolWrapper> for PoolState {
+    fn from(pool: PoolWrapper) -> Self {
+        pool.state
+    }
+}
+
+impl From<&PoolWrapper> for PoolState {
+    fn from(pool: &PoolWrapper) -> Self {
+        pool.state.clone()
+    }
+}
+
+impl From<PoolWrapper> for Vec<Replica> {
+    fn from(pool: PoolWrapper) -> Self {
+        pool.replicas
+    }
+}
+
+impl From<&PoolWrapper> for Vec<Replica> {
+    fn from(pool: &PoolWrapper) -> Self {
+        pool.replicas.clone()
+    }
+}
+
+// 1. state ( online > degraded )
+// 2. smaller n replicas
+// (here we should have pool IO stats over time so we can pick less active
+// pools rather than the number of replicas which is useless if the volumes
+// are not active)
+impl PartialOrd for PoolWrapper {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        match self.state.status.partial_cmp(&other.state.status) {
+            Some(Ordering::Greater) => Some(Ordering::Greater),
+            Some(Ordering::Less) => Some(Ordering::Less),
+            Some(Ordering::Equal) => match self.replicas.len().cmp(&other.replicas.len()) {
+                Ordering::Greater => Some(Ordering::Greater),
+                Ordering::Less => Some(Ordering::Less),
+                Ordering::Equal => Some(self.free_space().cmp(&other.free_space())),
+            },
+            None => None,
+        }
+    }
+}
+
+impl Ord for PoolWrapper {
+    fn cmp(&self, other: &Self) -> Ordering {
+        match self.state.status.partial_cmp(&other.state.status) {
+            Some(Ordering::Greater) => Ordering::Greater,
+            Some(Ordering::Less) => Ordering::Less,
+            Some(Ordering::Equal) => match self.replicas.len().cmp(&other.replicas.len()) {
+                Ordering::Greater => Ordering::Greater,
+                Ordering::Less => Ordering::Less,
+                Ordering::Equal => self.free_space().cmp(&other.free_space()),
+            },
+            None => Ordering::Equal,
+        }
+    }
+}
